backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
tensorflow==2.20.0
numpy==2.2.6
h5py==3.14.0
pillow==11.3.0
python-multipart==0.0.20
